Formalise the feature markers, calculate hit dice
Adding a more formal parsing of the feature markers for adding hit
dice, spell slots, etc.

Using the markers, collects the Hit Dice assigned at levels. Rolls the
die and adds hit points, as well as calculating the min and max scores
that could have been rolled. (Note, this is not a valid calculation in
5e as it takes no account of the character's Constitution modifier, nor
1st level always being max HP, but it is a starting point.)